+++ src/trust/nccgroup/decoderimproved/Utils.java
@@ -9,6 +9,7 @@ import org.exbin.utils.binary_data.BinaryData;
 
 import javax.swing.*;
 
+
 /**
  * Created by j on 10/26/16.
  */
