+++ src/trust/nccgroup/decoderimproved/modes/PrettifyMode.java
@@ -7,7 +7,6 @@ import trust.nccgroup.decoderimproved.modifiers.prettifiers.JsPrettifier;
 import trust.nccgroup.decoderimproved.modifiers.prettifiers.XmlPrettifier;
 
 import javax.swing.*;
-import java.awt.*;
 import java.util.ArrayList;
 
 public class PrettifyMode extends AbstractModificationMode {
